Added millions support to user count
Added support for millions, billions, and trillions to the user count.
Insignificant display change, but it was starting to bother me as it wasn't fixed. Previously, RoVer would display "Listening to 1043.4K users" as millions were not supported. Now, it displays "Listening to 1.0M users". Billions and trillions were also added, and although it's excessive, it's better safe than sorry.